+++ nptyping/ndarray.py
@@ -1,7 +1,7 @@
 """
 MIT License
 
-Copyright (c) 2022 Ramon Hagenaars
+Copyright (c) 2023 Ramon Hagenaars
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -21,6 +21,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
+import inspect
 from abc import ABC
 from typing import Any, Tuple
 
@@ -63,6 +64,10 @@ class NDArrayMeta(
 
     __args__: Tuple[Shape, DType]
     _parameterized: bool
+
+    @property
+    def __module__(cls) -> str:
+        return cls._get_module(inspect.stack(), "nptyping.ndarray")
 
     def _get_item(cls, item: Any) -> Tuple[Any, ...]:
         cls._check_item(item)